+++ tests/FreeElephants/JsonApiToolkit/OasToolsAdapter/YamlFileParserTest.php
@@ -10,7 +10,7 @@ class YamlFileParserTest extends AbstractTestCase
     public function testParse()
     {
         $parser = new YamlFileParser();
-        $openapi = $parser->parse(self::FIXTERE_PATH . '/petstore.yaml');
+        $openapi = $parser->parse(self::FIXTURE_PATH . '/petstore.yaml');
         $this->assertSame('Swagger Petstore', $openapi->info->title);
     }
 }