+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Data Transfer Object classes generation from swagger spec
 - JsonApiResponseFactory::createResponse() `meta` and `links` arguments 
 - CacheableDispatcherFactoryProxy
+- Union types support for `data.relationships` DTO: use suitable type for input data structure 
 
 ### Changed
 - Best composer dependencies compatibility: allow more versions